clean up button stories (#5991)
* clean up button stories

* clean up button stories

* put back snapshot
